sdk/tests/conformance/extensions/ext-frag-depth.html

<!--

/*
** Copyright (c) 2012 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebGL EXT_frag_depth Conformance Tests</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../resources/desktop-gl-constants.js" type="text/javascript"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/webgl-test.js"></script>
<script src="../resources/webgl-test-utils.js"></script>
</head>
<body>
<div id="description"></div>
<canvas id="canvas" style="width: 50px; height: 50px;"> </canvas>
<div id="console"></div>
<!-- Shaders for testing fragment depth writing -->

<!-- Shader omitting the required #extension pragma -->
<script id="missingPragmaFragmentShader" type="x-shader/x-fragment">
precision mediump float;
void main() {
    gl_FragColor = vec4(1.0, 0.0, 0.0, 1.0);
    gl_FragDepthEXT = 1.0;
}
</script>

<!-- Shader to test macro definition -->
<script id="macroFragmentShader" type="x-shader/x-fragment">
precision mediump float;
void main() {
#ifdef GL_EXT_frag_depth
    gl_FragColor = vec4(1.0, 0.0, 0.0, 1.0);
#else
    // Error expected
    #error no GL_EXT_frag_depth;
#endif
}
</script>

<!-- Shader with required #extension pragma -->
<script id="testFragmentShader" type="x-shader/x-fragment">
#extension GL_EXT_frag_depth : enable
precision mediump float;
void main() {
    gl_FragColor = vec4(1.0, 0.0, 0.0, 1.0);
    gl_FragDepthEXT = 1.0;
}
</script>
<!-- Shaders to link with test fragment shaders -->
<script id="goodVertexShader" type="x-shader/x-vertex">
attribute vec4 vPosition;
void main() {
    gl_Position = vPosition;
}
</script>
<!-- Shaders to test output -->
<script id="outputVertexShader" type="x-shader/x-vertex">
attribute vec4 vPosition;
varying vec4 position;
void main() {
    position = vPosition;
    gl_Position = vPosition;
}
</script>
<script id="outputFragmentShader" type="x-shader/x-fragment">
#extension GL_EXT_frag_depth : enable
precision mediump float;
varying vec4 position;
void main() {
    gl_FragColor = vec4(1.0, 0.0, 0.0, 1.0);
    gl_FragDepthEXT = position.z;
}
</script>

<script>
"use strict";
description("This test verifies the functionality of the EXT_frag_depth extension, if it is available.");

debug("");

var wtu = WebGLTestUtils;
var canvas = document.getElementById("canvas");
var gl = wtu.create3DContext(canvas);
var ext = null;

if (!gl) {
    testFailed("WebGL context does not exist");
} else {
    testPassed("WebGL context exists");

    runShaderTests(false);

    // Query the extension and store globally so shouldBe can access it
    ext = gl.getExtension("EXT_frag_depth");
    if (!ext) {
        testPassed("No EXT_frag_depth support -- this is legal");

        runSupportedTest(false);
    } else {
        testPassed("Successfully enabled EXT_frag_depth extension");

        runSupportedTest(true);

        runShaderTests(true);
        runOutputTests();
        runUniqueObjectTest();
    }
}

function runSupportedTest(extensionEnabled) {
    var supported = gl.getSupportedExtensions();
    if (supported.indexOf("EXT_frag_depth") >= 0) {
        if (extensionEnabled) {
            testPassed("EXT_frag_depth listed as supported and getExtension succeeded");
        } else {
            testFailed("EXT_frag_depth listed as supported but getExtension failed");
        }
    } else {
        if (extensionEnabled) {
            testFailed("EXT_frag_depth not listed as supported but getExtension succeeded");
        } else {
            testPassed("EXT_frag_depth not listed as supported and getExtension failed -- this is legal");
        }
    }
}

function runShaderTests(extensionEnabled) {
    debug("");
    debug("Testing various shader compiles with extension " + (extensionEnabled ? "enabled" : "disabled"));

    // Expect the macro shader to succeed ONLY if enabled
    var macroFragmentProgram = wtu.loadProgramFromScriptExpectError(gl, "goodVertexShader", "macroFragmentShader");
    if (extensionEnabled) {
        if (macroFragmentProgram) {
            // Expected result
            testPassed("GL_EXT_frag_depth defined in shaders when extension is enabled");
        } else {
            testFailed("GL_EXT_frag_depth not defined in shaders when extension is enabled");
        }
    } else {
        if (macroFragmentProgram) {
            testFailed("GL_EXT_frag_depth defined in shaders when extension is disabled");
        } else {
            testPassed("GL_EXT_frag_depth not defined in shaders when extension disabled");
        }
    }

    // Always expect the shader missing the #pragma to fail (whether enabled or not)
    var missingPragmaFragmentProgram = wtu.loadProgramFromScriptExpectError(gl, "goodVertexShader", "missingPragmaFragmentShader");
    if (missingPragmaFragmentProgram) {
        testFailed("Shader built-ins allowed without #extension pragma");
    } else {
        testPassed("Shader built-ins disallowed without #extension pragma");
    }

    // Try to compile a shader using the built-ins that should only succeed if enabled
    var testFragmentProgram = wtu.loadProgramFromScriptExpectError(gl, "goodVertexShader", "testFragmentShader");
    if (extensionEnabled) {
        if (testFragmentProgram) {
            testPassed("Shader built-ins compiled successfully when extension enabled");
        } else {
            testFailed("Shader built-ins failed to compile when extension enabled");
        }
    } else {
        if (testFragmentProgram) {
            testFailed("Shader built-ins compiled successfully when extension disabled");
        } else {
            testPassed("Shader built-ins failed to compile when extension disabled");
        }
    }
}

function runOutputTests() {
    // This tests does several draws with various values of z.
    // The output of the fragment shader is:
    // [dFdx(z), dFdy(z), fwidth(z), 1.0]
    // The expected math: (note the conversion to uint8)
    //    canvas.width = canvas.height = 50
    //    dFdx = totalChange.x / canvas.width  = 0.5 / 50.0 = 0.01
    //    dFdy = totalChange.y / canvas.height = 0.5 / 50.0 = 0.01
    //    fw = abs(dFdx + dFdy) = 0.01 + 0.01 = 0.02
    //    r = floor(dFdx * 40.0 * 255) = 102
    //    g = floor(dFdy * 40.0 * 255) = 102
    //    b = floor(fw * 40.0 * 255) = 204

    var e = 2; // Amount of variance to allow in result pixels - may need to be tweaked higher

    debug("Testing various draws for valid built-in function behavior");

    canvas.width = 50; canvas.height = 50;
    gl.viewport(0, 0, canvas.width, canvas.height);
    
    // Enable depth testing with a clearDepth of 0.5
    // This makes it so that fragments are only rendered when
    // gl_fragDepthEXT is < 0.5
    gl.clearDepth(0.5);
    gl.enable(gl.DEPTH_TEST);

    var positionLoc = 0;
    var texcoordLoc = 1;
    var program = wtu.setupProgram(gl, ["outputVertexShader", "outputFragmentShader"], ['vPosition', 'texCoord0'], [0, 1]);
    var quadParameters = wtu.setupUnitQuad(gl, 0, 1);

    function readLocation(x, y) {
        var pixels = new Uint8Array(1 * 1 * 4);
        var px = Math.floor(x * canvas.width);
        var py = Math.floor(y * canvas.height);
        gl.readPixels(px, py, 1, 1, gl.RGBA, gl.UNSIGNED_BYTE, pixels);
        return pixels;
    };
    function toString(arr) {
        var s = "[";
        for (var n = 0; n < arr.length; n++) {
            s += arr[n];
            if (n < arr.length - 1) {
                s += ", ";
            }
        }
        return s + "]";
    };
    function expectResult(targetList, successMessage, failureMessage) {
        var locations = [
            readLocation(0.1, 0.1),
            readLocation(0.9, 0.1),
            readLocation(0.1, 0.9),
            readLocation(0.9, 0.9),
        ];
        var anyDiffer = false;
        for (var n = 0; n < locations.length; n++) {
            var source = locations[n];
            var target = targetList[n];
            for (var m = 0; m < 4; m++) {
                if (Math.abs(source[m] - target[m]) > e) {
                    anyDiffer = true;
                    testFailed(failureMessage + "; Sample " + n + " should be " + toString(target) + ", was " + toString(source));
                    break;
                }
            }
        }
        if (!anyDiffer) {
            testPassed(successMessage);
        }
    };

    function setupBuffers(tl, tr, bl, br) {
        gl.bindBuffer(gl.ARRAY_BUFFER, quadParameters[0]);
        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array([
           1.0,  1.0, tr,
          -1.0,  1.0, tl,
          -1.0, -1.0, bl,
           1.0,  1.0, tr,
          -1.0, -1.0, bl,
           1.0, -1.0, br]), gl.STATIC_DRAW);
        gl.vertexAttribPointer(positionLoc, 3, gl.FLOAT, false, 0, 0);
    };

    // Draw 1: (no variation)
    setupBuffers(0.0, 0.0, 0.0, 0.0);
    wtu.clearAndDrawUnitQuad(gl);
    expectResult([
                    [255, 0, 0, 255],
                    [255, 0, 0, 255],
                    [255, 0, 0, 255],
                    [255, 0, 0, 255]
                 ],
                 "Draw 1 returned the correct data",
                 "Draw 1 returned incorrect data");

    // Draw 2: (variation in x)
    setupBuffers(1.0, 0.0, 1.0, 0.0);
    wtu.clearAndDrawUnitQuad(gl);
    expectResult([
                    [255, 255, 255, 255],
                    [255, 0, 0, 255],
                    [255, 255, 255, 255],
                    [255, 0, 0, 255]
                 ],
                 "Draw 2 returned the correct data",
                 "Draw 2 returned incorrect data");

    // Draw 3: (variation in y)
    setupBuffers(1.0, 1.0, 0.0, 0.0);
    wtu.clearAndDrawUnitQuad(gl);
    expectResult([
                    [255, 0, 0, 255],
                    [255, 0, 0, 255],
                    [255, 255, 255, 255],
                    [255, 255, 255, 255]
                 ],
                 "Draw 3 returned the correct data",
                 "Draw 3 returned incorrect data");

    // Draw 4: (variation in x & y)
    setupBuffers(1.0, 0.5, 0.5, 0.0);
    wtu.clearAndDrawUnitQuad(gl);
    expectResult([
                    [255, 255, 255, 255],
                    [255, 0, 0, 255],
                    [255, 255, 255, 255],
                    [255, 255, 255, 255]
                 ],
                 "Draw 4 returned the correct data",
                 "Draw 4 returned incorrect data");
}

function runUniqueObjectTest()
{
    debug("Testing that getExtension() returns the same object each time");
    gl.getExtension("EXT_frag_depth").myProperty = 2;
    gc();
    shouldBe('gl.getExtension("EXT_frag_depth").myProperty', '2');
}

debug("");
var successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>

</body>
</html>
